Updated database

+++ database3.h
@@ -5,6 +5,8 @@
 #include <ctime>   // для time()
 #include "sqlite3.h"
 #include <exception>
+#include <algorithm>
+
 class KeyValueStore {
 public:
     KeyValueStore(const std::string& dbName) {
@@ -15,6 +17,17 @@ public:
         
         const char* sqlCreateTable = "CREATE TABLE IF NOT EXISTS kv_store (key TEXT PRIMARY KEY, value TEXT);";
         executeSQL(sqlCreateTable);
+
+
+        std::string sqlSelectAll = "SELECT key FROM kv_store;";
+        sqlite3_stmt* stmt;
+        if (sqlite3_prepare_v2(db, sqlSelectAll.c_str(), -1, &stmt, nullptr) != SQLITE_OK) {
+            std::cerr << "Ошибка подготовки SQL запроса: " << sqlite3_errmsg(db) << std::endl;
+            throw (std::runtime_error("Ошибка подготовки SQL запроса"));
+        }
+
+        words = this->getAll();
+        sqlite3_finalize(stmt);
     }
 
     ~KeyValueStore() {
@@ -22,6 +35,9 @@ public:
     }
 
     void add(const std::string& key, const std::string& value) {
+        if (find(key, words)){
+            throw(std::runtime_error("This word already in dictionay"));
+        }
         std::string sqlInsert = "INSERT OR REPLACE INTO kv_store (key, value) VALUES (?, ?);";
         sqlite3_stmt* stmt;
         
@@ -41,6 +57,9 @@ public:
     }
 
     void remove(const std::string& key) {
+        if (not(find(key, words))){
+            throw(std::runtime_error("This word not in dictionary"));
+        }
         std::string sqlDelete = "DELETE FROM kv_store WHERE key = ?;";
         sqlite3_stmt* stmt;
 
@@ -60,8 +79,12 @@ public:
 
     
     std::string get(const std::string& key) {
+        if (not(find(key, words))){
+            throw std::runtime_error("This word not in dictionary");
+        } 
         std::string sqlSelect = "SELECT value FROM kv_store WHERE key = ?;";
         sqlite3_stmt* stmt;
+        
 
         if (sqlite3_prepare_v2(db, sqlSelect.c_str(), -1, &stmt, nullptr) != SQLITE_OK) {
             std::cerr << "Ошибка подготовки SQL запроса: " << sqlite3_errmsg(db) << std::endl;
@@ -117,11 +140,16 @@ public:
             std::cerr << "Ошибка подготовки SQL запроса: " << sqlite3_errmsg(db) << std::endl;
             throw std::runtime_error("Ошибка подготовки SQL запроса");
         }
-
+        int amount = count;
+        int k = 0;
         while (sqlite3_step(stmt) == SQLITE_ROW) {
+            if (k >= amount){
+                 break;
+            }
             const unsigned char* key = sqlite3_column_text(stmt, 0);
             const unsigned char* value = sqlite3_column_text(stmt, 1);
             pairs.emplace_back(reinterpret_cast<const char*>(key), reinterpret_cast<const char*>(value));
+            k += 1;
         }
         
         sqlite3_finalize(stmt);
@@ -130,9 +158,21 @@ public:
         return pairs;
     }
 
+    bool find(const std::string &word, std::vector<std::pair<std::string,std::string>>& words){
+        //auto it = std::find(words.begin(), words.end(), word);
+        for (int i = 0; i < words.size(); ++i){
+            if (words[i].first == word){
+                return true;
+            }
+        }
+        return false;
+
+        return 0;
+    }
+
 private:
     sqlite3* db;
-
+    std::vector<std::pair<std::string, std::string>> words;
     void executeSQL(const char* sql) {
         char* errMsg = nullptr;
         if (sqlite3_exec(db, sql, nullptr, nullptr, &errMsg) != SQLITE_OK) {
